Move the stable-diffusion modeling code so that it's easier to re-use.
