+++ candle-examples/examples/stable-diffusion/main.rs
@@ -4,17 +4,7 @@ extern crate accelerate_src;
 #[cfg(feature = "mkl")]
 extern crate intel_mkl_src;
 
-mod attention;
-mod clip;
-mod ddim;
-mod embeddings;
-mod resnet;
-mod schedulers;
-mod stable_diffusion;
-mod unet_2d;
-mod unet_2d_blocks;
-mod utils;
-mod vae;
+use candle_transformers::models::stable_diffusion;
 
 use anyhow::{Error as E, Result};
 use candle::{DType, Device, IndexOp, Tensor, D};
