+++ candle-transformers/src/models/mod.rs
@@ -6,4 +6,5 @@ pub mod falcon;
 pub mod llama;
 pub mod quantized_llama;
 pub mod segment_anything;
+pub mod stable_diffusion;
 pub mod whisper;
