+++ candle-transformers/src/models/stable_diffusion/attention.rs
