+++ candle-transformers/src/models/stable_diffusion/clip.rs
