+++ candle-transformers/src/models/stable_diffusion/embeddings.rs
