+++ candle-transformers/src/models/stable_diffusion/schedulers.rs
