+++ candle-transformers/src/models/stable_diffusion/utils.rs
